Updates Master Control to use React
This includes a nodenotes API to handle notes from editors
Also a node cloning and nuke interface, including modal confirmation
boxes

This also cleans up nodelet rendering from within the mid-tier Mason2
template system, which I'm working on winding down. It also removes a
fair bit of rendering from the system, and pushes it to the client. This
is all a major part of moving the system to be more modern and
mobile-first.

Fixes #3752